requirements-inference.txt
# RSNA 2019 Intracranial Hemorrhage Detection - Inference Dependencies
# Note: PyTorch and torchvision are provided by the base Docker image

# Pretrained model libraries
pretrainedmodels>=0.7.4
efficientnet-pytorch>=0.7.1

# Medical imaging
pydicom>=2.3.0

# Image processing
albumentations>=1.3.0
opencv-contrib-python>=4.5.0
scikit-image>=0.19.0
Pillow>=9.0.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.64.0
joblib>=1.1.0

# Google Drive download (for pretrained models)
gdown>=4.6.0

# Optional: For notebook support
ipykernel>=6.0.0
ipywidgets>=8.0.0
